refac(math): prevent symbol conflict `SupportsPolyXXX()`
